move prepocessing task tests into tasks/
